Basic Files from Kivy Planet
I was curious to see how the Facebook SDK could work on Kivy. Obviously,
it wouldn't be so simple as doing it in Android Dev Studio, since the
Facebook Android SDK is developed with Java.

I will soon be exploring the use of Pyjnius to implement a Java Class
using Python, and use Python to define a callback method.

The files were retrieved from:
https://github.com/tito/p4a-facebook-example
Basic Documentation:
http://kivy.org/planet/2013/08/using-facebook-sdk-with-python-for-android-kivy/

The documentation was somewhat sketchy. Time for an improved
documentation